1: Install react router 6
