+++ frontend/src/pages/Home/Home.tsx
@@ -1,0 +1,25 @@
+import logo from "../../logo.svg";
+import "./Home.scss";
+
+const Home = () => {
+	return (
+		<div className="App">
+			<header className="App-header">
+				<img src={logo} className="App-logo" alt="logo" />
+				<p>
+					Edit <code>src/App.tsx</code> and save to reload.
+				</p>
+				<a
+					className="App-link"
+					href="https://reactjs.org"
+					target="_blank"
+					rel="noopener noreferrer"
+				>
+					Learn React
+				</a>
+			</header>
+		</div>
+	);
+};
+
+export default Home;
